added comments for implicit and connected animations
