+++ Microsoft.Toolkit.Uwp.UI.Animations/Implicit.cs
@@ -11,38 +11,40 @@
 // ******************************************************************
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.Foundation.Metadata;
 using Windows.UI.Composition;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Hosting;
 
 namespace Microsoft.Toolkit.Uwp.UI.Animations
 {
+    /// <summary>
+    /// Attached Properties to enable Implicit Animations through XAML
+    /// </summary>
     public class Implicit
     {
+        /// <summary>
+        /// Identifies the Implicit.ShowAnimations XAML attached property
+        /// </summary>
         public static readonly DependencyProperty ShowAnimationsProperty =
-            DependencyProperty.RegisterAttached("ShowAnimations",
-                                                typeof(AnimationCollection),
-                                                typeof(Implicit),
-                                                new PropertyMetadata(null, ShowAnimationsChanged));
+            DependencyProperty.RegisterAttached("ShowAnimations", typeof(AnimationCollection), typeof(Implicit), new PropertyMetadata(null, ShowAnimationsChanged));
 
+        /// <summary>
+        /// Identifies the Implicit.HideAnimations XAML attached property
+        /// </summary>
         public static readonly DependencyProperty HideAnimationsProperty =
-            DependencyProperty.RegisterAttached("HideAnimations",
-                                                typeof(AnimationCollection),
-                                                typeof(Implicit),
-                                                new PropertyMetadata(null, HideAnimationsChanged));
+            DependencyProperty.RegisterAttached("HideAnimations", typeof(AnimationCollection), typeof(Implicit), new PropertyMetadata(null, HideAnimationsChanged));
 
-        // Using a DependencyProperty as the backing store for Animations.  This enables animation, styling, binding, etc...
+        /// <summary>
+        /// Identifies the Implicit.Animations XAML attached property
+        /// </summary>
         public static readonly DependencyProperty AnimationsProperty =
-            DependencyProperty.RegisterAttached("Animations",
-                                                typeof(AnimationCollection),
-                                                typeof(Implicit),
-                                                new PropertyMetadata(null, AnimationsChanged));
+            DependencyProperty.RegisterAttached("Animations", typeof(AnimationCollection), typeof(Implicit), new PropertyMetadata(null, AnimationsChanged));
 
+        /// <summary>
+        /// Gets the value of the Implicit.ShowAnimations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to get the value from</param>
+        /// <returns><see cref="AnimationCollection"/></returns>
         public static AnimationCollection GetShowAnimations(DependencyObject obj)
         {
             var collection = (AnimationCollection)obj.GetValue(ShowAnimationsProperty);
@@ -56,11 +58,21 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
             return collection;
         }
 
+        /// <summary>
+        /// Sets the value of the Implicit.ShowAnimations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to set the value</param>
+        /// <param name="value">The <see cref="AnimationCollection"/> to set</param>
         public static void SetShowAnimations(DependencyObject obj, AnimationCollection value)
         {
             obj.SetValue(ShowAnimationsProperty, value);
         }
 
+        /// <summary>
+        /// Gets the value of the Implicit.HideAnimations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to get the value from</param>
+        /// <returns><see cref="AnimationCollection"/></returns>
         public static AnimationCollection GetHideAnimations(DependencyObject obj)
         {
             var collection = (AnimationCollection)obj.GetValue(HideAnimationsProperty);
@@ -70,14 +82,25 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
                 collection = new AnimationCollection();
                 obj.SetValue(HideAnimationsProperty, collection);
             }
+
             return collection;
         }
 
+        /// <summary>
+        /// Sets the value of the Implicit.HideAnimations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to set the value</param>
+        /// <param name="value">The <see cref="AnimationCollection"/> to set</param>
         public static void SetHideAnimations(DependencyObject obj, AnimationCollection value)
         {
             obj.SetValue(HideAnimationsProperty, value);
         }
 
+        /// <summary>
+        /// Gets the value of the Implicit.Animations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to get the value from</param>
+        /// <returns><see cref="AnimationCollection"/></returns>
         public static AnimationCollection GetAnimations(DependencyObject obj)
         {
             var collection = (AnimationCollection)obj.GetValue(AnimationsProperty);
@@ -87,9 +110,15 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
                 collection = new AnimationCollection();
                 obj.SetValue(AnimationsProperty, collection);
             }
+
             return collection;
         }
 
+        /// <summary>
+        /// Sets the value of the Implicit.Animations XAML attached property.
+        /// </summary>
+        /// <param name="obj">The <see cref="FrameworkElement"/> to set the value</param>
+        /// <param name="value">The <see cref="AnimationCollection"/> to set</param>
         public static void SetAnimations(DependencyObject obj, AnimationCollection value)
         {
             obj.SetValue(AnimationsProperty, value);
@@ -107,21 +136,13 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
                 oldCollection.AnimationCollectionChanged -= ShowCollectionChanged;
             }
 
-            if (!(e.NewValue is AnimationCollection animationCollection))
+            if (e.NewValue is AnimationCollection animationCollection && d is UIElement element)
             {
-                return;
+                animationCollection.Element = element;
+                animationCollection.AnimationCollectionChanged -= ShowCollectionChanged;
+                animationCollection.AnimationCollectionChanged += ShowCollectionChanged;
+                ElementCompositionPreview.SetImplicitShowAnimation(element, GetCompositionAnimationGroup(animationCollection, element));
             }
-
-            if (!(d is UIElement element))
-            {
-                return;
-            }
-
-            animationCollection.Element = element;
-            animationCollection.AnimationCollectionChanged -= ShowCollectionChanged;
-            animationCollection.AnimationCollectionChanged += ShowCollectionChanged;
-
-            ElementCompositionPreview.SetImplicitShowAnimation(element, GetCompositionAnimationGroup(animationCollection, element));
         }
 
         private static void HideAnimationsChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
@@ -136,21 +157,13 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
                 oldCollection.AnimationCollectionChanged -= HideCollectionChanged;
             }
 
-            if (!(e.NewValue is AnimationCollection animationCollection))
+            if (e.NewValue is AnimationCollection animationCollection && d is UIElement element)
             {
-                return;
+                animationCollection.Element = element;
+                animationCollection.AnimationCollectionChanged -= HideCollectionChanged;
+                animationCollection.AnimationCollectionChanged += HideCollectionChanged;
+                ElementCompositionPreview.SetImplicitHideAnimation(element, GetCompositionAnimationGroup(animationCollection, element));
             }
-
-            if (!(d is UIElement element))
-            {
-                return;
-            }
-
-            animationCollection.Element = element;
-            animationCollection.AnimationCollectionChanged -= HideCollectionChanged;
-            animationCollection.AnimationCollectionChanged += HideCollectionChanged;
-
-            ElementCompositionPreview.SetImplicitHideAnimation(element, GetCompositionAnimationGroup(animationCollection, element));
         }
 
         private static void AnimationsChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
@@ -160,21 +173,13 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
                 oldCollection.AnimationCollectionChanged -= AnimationsCollectionChanged;
             }
 
-            if (!(e.NewValue is AnimationCollection animationCollection))
+            if (e.NewValue is AnimationCollection animationCollection && d is UIElement element)
             {
-                return;
+                animationCollection.Element = element;
+                animationCollection.AnimationCollectionChanged -= AnimationsCollectionChanged;
+                animationCollection.AnimationCollectionChanged += AnimationsCollectionChanged;
+                ElementCompositionPreview.GetElementVisual(element).ImplicitAnimations = GetImplicitAnimationCollection(animationCollection, element);
             }
-
-            if (!(d is UIElement element))
-            {
-                return;
-            }
-
-            animationCollection.Element = element;
-            animationCollection.AnimationCollectionChanged -= AnimationsCollectionChanged;
-            animationCollection.AnimationCollectionChanged += AnimationsCollectionChanged;
-
-            ElementCompositionPreview.GetElementVisual(element).ImplicitAnimations = GetImplicitAnimationCollection(animationCollection, element);
         }
 
         private static void ShowCollectionChanged(object sender, EventArgs e)
@@ -207,7 +212,6 @@ namespace Microsoft.Toolkit.Uwp.UI.Animations
             }
 
             ElementCompositionPreview.SetImplicitHideAnimation(collection.Element, GetCompositionAnimationGroup(collection, collection.Element));
-
         }
 
         private static void AnimationsCollectionChanged(object sender, EventArgs e)
